factor out upsert tests to their own file
